Refactor centroid code into util
